requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
dj-rest-auth==6.0.0
django-allauth==65.0.2
django-filter==24.3
drf-spectacular==0.27.2
ruff==0.6.9
gunicorn==23.0.0
Pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2024.2
redis==5.1.1
sqlparse==0.5.1
uvicorn==0.31.1